File Organization, Title screen, Transitions, MainMenu,, and Fonts
This commit adds a basic title screen (NOT FINAL) and adds a fade in and out effect when switching scenes.  Also adds a main Menu Scene (It has nothing yet) and organized a few files.
Added a (maybe) temporary font called "Rumble Tumble", further use of it should be discussed.