Grappling attack (SKELETON)
NOTE: THIS IS A MODIFIABLE ATTACK THIS IS JUST THE SKELETON.
You can initiate the grapple with 'G' and the "bot" will take down the grapple meter by 1 every second.  The player can increase the meter by pressing space.  -10 grapple meter means the "bot" won and 10 means the player won (Viewable in the console).  There are no animations for the grapple attacks and meter yet.