Pathfinding direction hint line
Adds a new script which creates a line from one point to the other using A* pathfinding.  Should be used to to direct the player where to go.  Removed the LineToCoachScript and is now replaced with the DirectionalHintPathfind.  Added two test objects for messing around with this system and more.  Both are under the parent class called TestObject and is set as "Inactive".